Module libra_crypto::ed25519::compat

Those transitory traits are meant to help with the progressive migration of the code base to the crypto module and will disappear after

Functions

generate_genesis_keypair
Generates a well-known keypair (Ed25519PrivateKey, Ed25519PublicKey) for special use in the genesis block. A genesis block is the first block of a blockchain and it is hardcoded as it's a special case in that it does not reference a previous block.

generate_keypair
Generate an arbitrary key pair, with possible Rng input

keypair_strategy
Used to produce keypairs from a seed for testing purposes