Module libra_crypto::slip0010

This module provides an API for SLIP-0010 and for the Ed25519 curve based on SLIP-0010 : Universal private key derivation from master private key.

SLIP-0010 describes how to derive a master private/public key for various curves and how a BIP-0032 like derivation is used to hierarchically generate keys from an original seed.

Note that SLIP-0010 only supports private parent key → private child key hardened key generation for Ed25519. This key derivation protocol should be preferred to HKDF if compatibility with hardware wallets and the BIP32 protocol is required.

Structs

ExtendedPrivKey
Extended private key that includes additional child_path and chain-code.

Slip0010
SLIP-0010 structure.

Enums

Slip0010Error
An error type for SLIP-0010 key derivation issues.