Module libra_crypto::traits

This module provides a generic set of traits for dealing with cryptographic primitives.

For examples on how to use these traits, see the implementations of the [ed25519] or [bls12381] modules.

Enums

CryptoMaterialError
An error type for key and signature validation issues, see [ValidKey][ValidKey].

Traits

ExchangeKey
A type family for Diffie-Hellman private key material

Genesis
A type family with a by-convention notion of genesis private key.

PrivateKey
A type family for key material that should remain secret and has an associated type of the [PublicKey][PublicKey] family.

PublicKey
A type for key material that can be publicly shared, and in asymmetric fashion, can be obtained from a [PrivateKey][PrivateKey] reference. This convertibility requirement ensures the existence of a deterministic, canonical public key construction from a private key.

Signature
A type family for signature material that knows which public key type is needed to verify it, and given such a public key, knows how to verify.

SigningKey
A type family of valid keys that know how to sign.

Uniform
A type family for schemes which know how to generate key material from a cryptographically-secure [CryptoRng][::rand::CryptoRng].

ValidKey
Key material with a notion of byte validation.

ValidKeyStringExt
An extension to to/from Strings for [ValidKey][ValidKey].

VerifyingKey
A type family of public keys that are used for signing.