Trait libra_crypto::traits::Signature

pub trait Signature: for<'a> TryFrom<&'a [u8], Error = CryptoMaterialError> + Sized + Debug + Clone + Eq + Hash + Sealed {
    type VerifyingKeyMaterial: VerifyingKey<SignatureMaterial = Self>;
    type SigningKeyMaterial: SigningKey<SignatureMaterial = Self>;
    fn verify(
        &self, 
        message: &HashValue, 
        public_key: &Self::VerifyingKeyMaterial
    ) -> Result<()>;

    fn verify_arbitrary_msg(
        &self, 
        message: &[u8], 
        public_key: &Self::VerifyingKeyMaterial
    ) -> Result<()>;

    fn to_bytes(&self) -> Vec<u8>;

    fn batch_verify_signatures(
        message: &HashValue, 
        keys_and_signatures: Vec<(Self::VerifyingKeyMaterial, Self)>
    ) -> Result<()> { ... }
}

A type family for signature material that knows which public key type is needed to verify it, and given such a public key, knows how to verify.

This trait simply requires an association to some type of the PublicKey family of which we are the SignatureMaterial.

This trait has a requirement on a pub(crate) marker trait meant to specifically limit its implementations to the present crate.

It should be possible to write a generic signature function that checks signature material passed as &[u8] and only returns Ok when that material de-serializes to a signature of the expected concrete scheme. This would be done as an extension trait of Signature.

Associated Types

type VerifyingKeyMaterial: VerifyingKey<SignatureMaterial = Self>

The associated verifying key type for this signature.

type SigningKeyMaterial: SigningKey<SignatureMaterial = Self>

The associated signing key type for this signature

Loading content...
Required methods

fn verify(
&self,
message: &HashValue,
public_key: &Self::VerifyingKeyMaterial
) -> Result<()>

The verification function.

fn verify_arbitrary_msg(
&self,
message: &[u8],
public_key: &Self::VerifyingKeyMaterial
) -> Result<()>

Native verification function.

fn to_bytes(&self) -> Vec<u8>

Convert the signature into a byte representation.

Loading content...
Provided methods

fn batch_verify_signatures(
message: &HashValue,
keys_and_signatures: Vec<(Self::VerifyingKeyMaterial, Self)>
) -> Result<()>

The implementer can override a batch verification implementation that by default iterates over each signature. More efficient implementations exist and should be implemented for many schemes.

Loading content...