Trait libra_crypto::traits::SigningKey

pub trait SigningKey: PrivateKey<PublicKeyMaterial = Self::VerifyingKeyMaterial> + ValidCryptoMaterial + Sealed {
    type VerifyingKeyMaterial: VerifyingKey<SigningKeyMaterial = Self>;
    type SignatureMaterial: Signature<SigningKeyMaterial = Self>;
    fn sign_message(&self, message: &HashValue) -> Self::SignatureMaterial;

    fn sign_arbitrary_message(&self, message: &[u8]) -> Self::SignatureMaterial;

    fn verifying_key(&self) -> Self::VerifyingKeyMaterial { ... }
}

A type family of valid keys that know how to sign.

This trait has a requirement on a pub(crate) marker trait meant to specifically limit its implementations to the present crate.

A trait for a ValidCryptoMaterial which knows how to sign a message, and return an associated Signature type.

Associated Types

type VerifyingKeyMaterial: VerifyingKey<SigningKeyMaterial = Self>

The associated verifying key type for this signing key.

type SignatureMaterial: Signature<SigningKeyMaterial = Self>

The associated signature type for this signing key.

Loading content...
Required methods

fn sign_message(&self, message: &HashValue) -> Self::SignatureMaterial

Signs an input message.

fn sign_arbitrary_message(&self, message: &[u8]) -> Self::SignatureMaterial

Signs a non-hash input message. For testing only.

Loading content...
Provided methods

fn verifying_key(&self) -> Self::VerifyingKeyMaterial

Returns the associated verifying key

Loading content...