Trait libra_crypto::traits::VerifyingKey

pub trait VerifyingKey: PublicKey<PrivateKeyMaterial = Self::SigningKeyMaterial> + ValidKey {
    type SigningKeyMaterial: SigningKey<VerifyingKeyMaterial = Self>;
    type SignatureMaterial: Signature<VerifyingKeyMaterial = Self>;
    fn verify_signature(
        &self, 
        message: &HashValue, 
        signature: &Self::SignatureMaterial
    ) -> Result<()> { ... }

    fn batch_verify_signatures(
        message: &HashValue, 
        keys_and_signatures: Vec<(Self, Self::SignatureMaterial)>
    ) -> Result<()> { ... }
}

A type family of public keys that are used for signing.

It is linked to a type of the Signature family, which carries the verification implementation.

Associated Types

type SigningKeyMaterial: SigningKey<VerifyingKeyMaterial = Self>

The associated signing key type for this verifying key.

type SignatureMaterial: Signature<VerifyingKeyMaterial = Self>

The associated signature type for this verifying key.

Loading content...
Provided methods

fn verify_signature(
&self,
message: &HashValue,
signature: &Self::SignatureMaterial
) -> Result<()>

We provide the logical implementation which dispatches to the signature.

fn batch_verify_signatures(
message: &HashValue,
keys_and_signatures: Vec<(Self, Self::SignatureMaterial)>
) -> Result<()>

We provide the implementation which dispatches to the signature.

Loading content...