Trait libra_crypto::hash::CryptoHasher

pub trait CryptoHasher: Default {
    fn finish(self) -> HashValue;

    fn write(&mut self, bytes: &[u8]) -> &mut Self;

    fn write_u8(&mut self, byte: u8) { ... }
}

A trait for generating hash from arbitrary stream of bytes.

Instances of CryptoHasher usually represent state that is changed while hashing data. Similar to std::hash::Hasher but not same. CryptoHasher cannot be reused after finish() has been called.

Required methods

fn finish(self) -> HashValue

Finish constructing the HashValue.

fn write(&mut self, bytes: &[u8]) -> &mut Self

Write bytes into the hasher.

Loading content...
Provided methods

fn write_u8(&mut self, byte: u8)

Write a single byte into the hasher.

Loading content...