Trait libra_crypto::traits::PrivateKey

pub trait PrivateKey: Sized {
    type PublicKeyMaterial: PublicKey<PrivateKeyMaterial = Self>;
    fn public_key(&self) -> Self::PublicKeyMaterial { ... }
}

A type family for key material that should remain secret and has an associated type of the PublicKey family.

Associated Types

type PublicKeyMaterial: PublicKey<PrivateKeyMaterial = Self>

We require public / private types to be coupled, i.e. their associated type is each other.

Loading content...
Provided methods

fn public_key(&self) -> Self::PublicKeyMaterial

Returns the associated public key

Loading content...